Update handling of BT.470 BG color transfer
ffmpeg does not support a passing a bt470bg value for color_trc.  Based on https://ffmpeg.org/ffmpeg-codecs.html, the equivalent value is gamma28.  This update maps the bt470bg value to gamma28 when setting color_trc.

There may be other similar values that need to be mapped but this is the only one I've encountered in my own testing.